Avoid GCC signed/unsigned warning by explicit cast
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@8620 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/ncbi_conn_streambuf.cpp
@@ -204,7 +204,7 @@ streamsize CConn_Streambuf::xsgetn(CT_CHAR_TYPE* buf, streamsize m)
             memcpy(buf, m_ReadBuf, x_read*sizeof(CT_CHAR_TYPE));
             setg(m_ReadBuf, m_ReadBuf + x_read, m_ReadBuf + xx_read);
         } else {
-            size_t xx_read = x_read > m_BufSize ? m_BufSize : x_read;
+            size_t xx_read = x_read > (size_t) m_BufSize ? m_BufSize : x_read;
             memcpy(m_ReadBuf,buf+x_read-xx_read,xx_read*sizeof(CT_CHAR_TYPE));
             setg(m_ReadBuf, m_ReadBuf + xx_read, m_ReadBuf + xx_read);
         }
@@ -269,6 +269,9 @@ END_NCBI_SCOPE
 /*
  * ---------------------------------------------------------------------------
  * $Log$
+ * Revision 6.27  2003/04/25 15:20:37  lavr
+ * Avoid GCC signed/unsigned warning by explicit cast
+ *
  * Revision 6.26  2003/04/11 17:57:11  lavr
  * Define xsgetn() unconditionally
  *
